+++ Coverless/Coverless/DesignSystem/Protocols/DiscoverViewDelegate.swift
@@ -1,0 +1,10 @@
+//
+//  DiscoverViewDelegate.swift
+//  Coverless
+//
+//  Created by Gabriel Ferreira de Carvalho on 01/10/21.
+//
+
+protocol DiscoverViewFooterDelegate: AnyObject {
+    func refreshList()
+}
