+++ Coverless/Coverless/Views/DiscoverView.swift
@@ -20,7 +20,8 @@ final class DiscoverView: UIView, Designable {
         let header = NSCollectionLayoutBoundarySupplementaryItem(
             layoutSize: .init(widthDimension: .fractionalWidth(1), heightDimension: .estimated(80)),
             elementKind: UICollectionView.elementKindSectionHeader, alignment: .top)
-        section.boundarySupplementaryItems = [header]
+        let footer = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: .init(widthDimension: .fractionalWidth(1), heightDimension: .estimated(55)), elementKind: UICollectionView.elementKindSectionFooter, alignment: .bottom)
+        section.boundarySupplementaryItems = [header, footer]
         //section.contentInsets = NSDirectionalEdgeInsets(top: 0.0, leading: 18.0, bottom: 0.0, trailing: 18.0)
         let layout = UICollectionViewCompositionalLayout(section: section)
         
@@ -32,6 +33,9 @@ final class DiscoverView: UIView, Designable {
         collectionView.register(SubjectCollectionHeader.self,
                                 forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
                                 withReuseIdentifier: SubjectCollectionHeader.identifier)
+        collectionView.register(DiscoverViewFooter.self,
+                                forSupplementaryViewOfKind: UICollectionView.elementKindSectionFooter,
+                                withReuseIdentifier: DiscoverViewFooter.identifier)
         
         collectionView.translatesAutoresizingMaskIntoConstraints = false
         return collectionView
